refactor : user agora totalmente funcional como antes
